requirements.txt
beautifulsoup4
google
google-api-python-client
google_auth_oauthlib
httpx
importlib_metadata
jinja2
markupsafe
openai
Pillow
pypdf
python-dateutil
python-docx
python-dotenv
PyYAML
requests
striprtf
truststore
